Author: glyph

Reviewer: ashfall

Fixes: #5734

Fix a minor defect in `_sigchld.c`; store the result of a function in a
container declared to be of the correct size, rather than the incorrect size.



git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@34829 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb

+++ twisted/internet/_sigchld.c
@@ -12,7 +12,7 @@ static int sigchld_pipe_fd = -1;
 
 static void got_signal(int sig) {
     int saved_errno = errno;
-    int ignored_result;
+    ssize_t ignored_result;
 
     /* write() errors are unhandled.  If the buffer is full, we don't
      * care.  What about other errors? */
